moved body component to typography
