+++ pages/index.js
@@ -12,7 +12,7 @@ import { getFavouritesUrl } from "../selectors/urls";
 import { globalTheme } from "../theme";
 import { css } from "react-emotion";
 import Container from "../components/container";
-import Body from "../components/body";
+import Body from "../components/typography/body";
 import Header1 from "../components/typography/header1";
 import Router from "next/router";
