+++ pages/map.js
@@ -13,7 +13,7 @@ import { globalTheme } from "../theme";
 import { css } from "react-emotion";
 import Container from "../components/container";
 import Header1 from "../components/typography/header1";
-import Body from "../components/body";
+import Body from "../components/typography/body";
 import HeaderButton from "../components/header_button";
 
 const paper = css`
